@@ -1,5 +1,7 @@
 export 'src/slide_countdown.dart';
 export 'src/slide_countdown_separated.dart';
-export 'src/utils/duration_title.dart';
+export 'src/widgets/raw_digit_item.dart';
+export 'src/widgets/raw_slide_countdown.dart';
+export 'src/models/duration_title.dart';
 export 'src/utils/enum.dart';
 export 'package:stream_duration/stream_duration.dart';
